refactor: move all const to seperate mock file
